Commit 0.2
In this change, I personalised the code that Linus and I worked on
together to fit into the context of my game. I also added a circle that
moves up, down, left and right. Next step is to get rid of the trailing
circles and when the circle collides with the other circles then the
other circles change color.