Commit 0.3
In this commit I worked on making the collision and making the circle
turn white when collided with. I successfully was able to program the
collision, and I was also able to turn one circle white when collided
with. However, I was not able to turn multiple circles white. This is
what I will be working on for next time.